Test bidService

+++ backend/src/main/java/es/udc/paproject/backend/model/services/BidService.java
@@ -1,12 +1,13 @@
 package es.udc.paproject.backend.model.services;
 
+import java.math.BigDecimal;
+
 import es.udc.paproject.backend.model.common.exceptions.InstanceNotFoundException;
 import es.udc.paproject.backend.model.entities.Bid;
-import es.udc.paproject.backend.model.entities.Product;
 
 public interface BidService {
 	
-	Bid createBid(Long id, Long productId, Float quantity) throws ExpiratedProductDateException, InstanceNotFoundException;
+	Bid createBid(Long id, Long productId, BigDecimal quantity) throws ExpiratedProductDateException, InstanceNotFoundException;
 	
 	Block<Bid> getUserBids(Long userId) throws InstanceNotFoundException;
